Code improvements

+++ app/src/main/java/mega/privacy/android/app/modalbottomsheet/FileContactsListBottomSheetDialogFragment.java
@@ -6,22 +6,15 @@ import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.Paint;
 import android.os.Bundle;
 import android.support.design.widget.BottomSheetBehavior;
 import android.support.design.widget.BottomSheetDialogFragment;
-import android.support.v4.content.ContextCompat;
 import android.util.DisplayMetrics;
 import android.view.Display;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-
 import java.io.File;
-import java.util.Locale;
-
 import mega.privacy.android.app.DatabaseHandler;
 import mega.privacy.android.app.MegaApplication;
 import mega.privacy.android.app.MegaContactDB;
@@ -60,7 +53,6 @@ public class FileContactsListBottomSheetDialogFragment extends BottomSheetDialog
     public EmojiTextView titleNameContactPanel;
     public TextView titleMailContactPanel;
     public RoundedImageView contactImageView;
-    public EmojiTextView avatarInitialLetter;
     public LinearLayout optionChangePermissions;
     public LinearLayout optionDelete;
     public LinearLayout optionInfo;
@@ -126,7 +118,6 @@ public class FileContactsListBottomSheetDialogFragment extends BottomSheetDialog
         titleNameContactPanel = contentView.findViewById(R.id.file_contact_list_contact_name_text);
         titleMailContactPanel = (TextView) contentView.findViewById(R.id.file_contact_list_contact_mail_text);
         contactImageView = (RoundedImageView) contentView.findViewById(R.id.sliding_file_contact_list_thumbnail);
-        avatarInitialLetter = contentView.findViewById(R.id.sliding_file_contact_list_initial_letter);
 
         optionChangePermissions = (LinearLayout) contentView.findViewById(R.id.file_contact_list_option_permissions_layout);
         optionDelete = (LinearLayout) contentView.findViewById(R.id.file_contact_list_option_delete_layout);
@@ -236,75 +227,27 @@ public class FileContactsListBottomSheetDialogFragment extends BottomSheetDialog
 
     public void addAvatarContactPanel(MegaUser contact){
 
+        /*Default Avatar*/
+        int color = colorAvatar(context, megaApi, contact);
+        contactImageView.setImageBitmap(getDefaultAvatar(context, color, fullName, AVATAR_SIZE, false));
+
+        /*Avatar*/
         if(contact!=null){
             String contactMail = contact.getEmail();
             File avatar = buildAvatarFile(getActivity(),contactMail + ".jpg");
             Bitmap bitmap = null;
-            if (isFileAvailable(avatar)){
-                if (avatar.length() > 0){
-                    BitmapFactory.Options bOpts = new BitmapFactory.Options();
-                    bOpts.inPurgeable = true;
-                    bOpts.inInputShareable = true;
-                    bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-                    if (bitmap == null) {
-                        avatar.delete();
-                    }
-                    else{
-                        avatarInitialLetter.setVisibility(View.GONE);
-                        contactImageView.setImageBitmap(bitmap);
-                        return;
-                    }
+            if (isFileAvailable(avatar) && avatar.length() > 0){
+                BitmapFactory.Options bOpts = new BitmapFactory.Options();
+                bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
+                if (bitmap == null) {
+                    avatar.delete();
+                }
+                else{
+                    contactImageView.setImageBitmap(bitmap);
+                    return;
                 }
             }
         }
-
-        ////DEfault AVATAR
-        Bitmap defaultAvatar = Bitmap.createBitmap(DEFAULT_AVATAR_WIDTH_HEIGHT, DEFAULT_AVATAR_WIDTH_HEIGHT, Bitmap.Config.ARGB_8888);
-        Canvas c = new Canvas(defaultAvatar);
-        Paint p = new Paint();
-        p.setAntiAlias(true);
-
-        if (contact != null) {
-            String color = megaApi.getUserAvatarColor(contact);
-            if (color != null) {
-                logDebug("The color to set the avatar is " + color);
-                p.setColor(Color.parseColor(color));
-            } else {
-                logDebug("Default color to the avatar");
-                p.setColor(ContextCompat.getColor(context, R.color.lollipop_primary_color));
-            }
-        } else {
-            logWarning("Contact is NULL");
-            p.setColor(ContextCompat.getColor(context, R.color.lollipop_primary_color));
-        }
-
-        int radius;
-        if (defaultAvatar.getWidth() < defaultAvatar.getHeight())
-            radius = defaultAvatar.getWidth() / 2;
-        else
-            radius = defaultAvatar.getHeight() / 2;
-
-        c.drawCircle(defaultAvatar.getWidth() / 2, defaultAvatar.getHeight() / 2, radius, p);
-        contactImageView.setImageBitmap(defaultAvatar);
-
-        Display display = getActivity().getWindowManager().getDefaultDisplay();
-        outMetrics = new DisplayMetrics();
-        display.getMetrics(outMetrics);
-
-        if(fullName == null || fullName.length() <= 0 || fullName.trim().length() <= 0){
-            avatarInitialLetter.setVisibility(View.INVISIBLE);
-        }else{
-            String firstLetter = getFirstLetter(fullName);
-            if(firstLetter.trim().isEmpty() || firstLetter.equals("(")){
-                avatarInitialLetter.setVisibility(View.INVISIBLE);
-            }else {
-                avatarInitialLetter.setText(firstLetter);
-                avatarInitialLetter.setTextColor(Color.WHITE);
-                avatarInitialLetter.setVisibility(View.VISIBLE);
-                avatarInitialLetter.setTextSize(22);
-            }
-        }
-
     }
 
     @Override
